Refactor Chart component
