+++ src/Chart/renderLabels.tsx
@@ -1,84 +1,13 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import Path from '../Path';
+import React from 'react';
 import { default as DefaultLabel, Props as LabelProps } from '../Label';
 import {
   degreesToRadians,
   evaluateLabelTextAnchor,
   extractPercentage,
-  valueBetween,
+  extractAbsoluteCoordinates,
 } from '../utils';
-import { dataPropType, stylePropType } from '../propTypes';
-import {
-  Data,
-  ExtendedData,
-  EventHandler,
-  LabelProp,
-  StyleObject,
-} from '../commonTypes';
-
-function extractAbsoluteCoordinates(props: Props) {
-  const [viewBoxWidth, viewBoxHeight] = props.viewBoxSize;
-  return {
-    cx: extractPercentage(props.cx, viewBoxWidth),
-    cy: extractPercentage(props.cy, viewBoxHeight),
-    radius: extractPercentage(props.radius, viewBoxWidth),
-  };
-}
-
-function sumValues(data: Data) {
-  return data.reduce((acc, dataEntry) => acc + dataEntry.value, 0);
-}
-
-// Append "percentage", "degrees" and "startOffset" into each data entry
-function extendData({
-  data,
-  lengthAngle: totalAngle,
-  totalValue,
-  paddingAngle,
-}: Props) {
-  const total = totalValue || sumValues(data);
-  const normalizedTotalAngle = valueBetween(totalAngle, -360, 360);
-  const numberOfPaddings =
-    Math.abs(normalizedTotalAngle) === 360 ? data.length : data.length - 1;
-  const singlePaddingDegrees = Math.abs(paddingAngle) * Math.sign(totalAngle);
-  const degreesTakenByPadding = singlePaddingDegrees * numberOfPaddings;
-  const degreesTakenByPaths = normalizedTotalAngle - degreesTakenByPadding;
-  let lastSegmentEnd = 0;
-
-  // @NOTE: Shall we evaluate percentage accordingly to dataEntry.value's sign?
-  return data.map(dataEntry => {
-    const valueInPercentage = total === 0 ? 0 : (dataEntry.value / total) * 100;
-    const degrees = extractPercentage(degreesTakenByPaths, valueInPercentage);
-    const startOffset = lastSegmentEnd;
-    lastSegmentEnd = lastSegmentEnd + degrees + singlePaddingDegrees;
-
-    return {
-      percentage: valueInPercentage,
-      degrees,
-      startOffset,
-      ...dataEntry,
-    };
-  });
-}
-
-function makeSegmentTransitionStyle(
-  duration: number,
-  easing: string,
-  furtherStyles: StyleObject = {}
-) {
-  // Merge CSS transition necessary for chart animation with the ones provided by "segmentsStyle"
-  const transition = [
-    `stroke-dashoffset ${duration}ms ${easing}`,
-    furtherStyles.transition,
-  ]
-    .filter(Boolean)
-    .join(',');
-
-  return {
-    transition,
-  };
-}
+import { ExtendedData, LabelProp } from '../commonTypes';
+import { Props as ChartProps } from './index';
 
 function renderLabelItem(
   option: LabelProp,
@@ -100,7 +29,7 @@ function renderLabelItem(
   return <DefaultLabel {...labelProps}>{label}</DefaultLabel>;
 }
 
-function renderLabels(data: ExtendedData, props: Props) {
+export default function renderLabels(data: ExtendedData, props: ChartProps) {
   const { cx, cy, radius } = extractAbsoluteCoordinates(props);
   const labelPosition = extractPercentage(radius, props.labelPosition);
 
@@ -131,231 +60,4 @@ function renderLabels(data: ExtendedData, props: Props) {
 
     return renderLabelItem(props.label, labelProps, dataEntry.value);
   });
-}
-
-function renderSegments(data: ExtendedData, props: Props, hide: Boolean) {
-  let style = props.segmentsStyle;
-  if (props.animate) {
-    const transitionStyle = makeSegmentTransitionStyle(
-      props.animationDuration,
-      props.animationEasing,
-      style
-    );
-    style = Object.assign({}, style, transitionStyle);
-  }
-
-  // Hide/reveal the segment?
-  let reveal: number;
-  if (hide === true) {
-    reveal = 0;
-  } else if (typeof props.reveal === 'number') {
-    reveal = props.reveal;
-  } else if (hide === false) {
-    reveal = 100;
-  }
-
-  const { cx, cy, radius } = extractAbsoluteCoordinates(props);
-  const lineWidth = extractPercentage(radius, props.lineWidth);
-  const paths = data.map((dataEntry, index) => {
-    const startAngle = props.startAngle + dataEntry.startOffset;
-
-    return (
-      <Path
-        key={dataEntry.key || index}
-        cx={cx}
-        cy={cy}
-        startAngle={startAngle}
-        lengthAngle={dataEntry.degrees}
-        radius={radius}
-        lineWidth={lineWidth}
-        reveal={reveal}
-        title={dataEntry.title}
-        style={Object.assign({}, style, dataEntry.style)}
-        stroke={dataEntry.color}
-        strokeLinecap={props.rounded ? 'round' : undefined}
-        fill="none"
-        onMouseOver={
-          props.onMouseOver &&
-          // @ts-ignore
-          ((e: React.MouseEvent) => props.onMouseOver(e, props.data, index))
-        }
-        onMouseOut={
-          props.onMouseOut &&
-          // @ts-ignore
-          ((e: React.MouseEvent) => props.onMouseOut(e, props.data, index))
-        }
-        onClick={
-          props.onClick &&
-          // @ts-ignore
-          ((e: React.MouseEvent) => props.onClick(e, props.data, index))
-        }
-      />
-    );
-  });
-
-  if (props.background) {
-    paths.unshift(
-      <Path
-        key="bg"
-        cx={cx}
-        cy={cy}
-        startAngle={props.startAngle}
-        lengthAngle={props.lengthAngle}
-        radius={radius}
-        lineWidth={lineWidth}
-        stroke={props.background}
-        strokeLinecap={props.rounded ? 'round' : undefined}
-        fill="none"
-      />
-    );
-  }
-
-  return paths;
-}
-
-type Props = typeof ReactMinimalPieChart.defaultProps & {
-  animate?: boolean;
-  animationDuration?: number;
-  animationEasing?: string;
-  background?: string;
-  className?: string;
-  cx?: number;
-  cy?: number;
-  data: Data;
-  lengthAngle?: number;
-  lineWidth?: number;
-  injectSvg?: () => React.ReactElement | void;
-  label?: LabelProp;
-  labelPosition?: number;
-  labelStyle?: StyleObject;
-  onClick?: EventHandler;
-  onMouseOut?: EventHandler;
-  onMouseOver?: EventHandler;
-  paddingAngle?: number;
-  radius?: number;
-  reveal?: number;
-  rounded?: boolean;
-  segmentsStyle?: StyleObject;
-  startAngle?: number;
-  style?: StyleObject;
-  totalValue?: number;
-  viewBoxSize?: [number, number];
-};
-
-export default class ReactMinimalPieChart extends Component<Props> {
-  static displayName = 'ReactMinimalPieChart';
-  static defaultProps = {
-    animate: false,
-    animationDuration: 500,
-    animationEasing: 'ease-out',
-    cx: 50,
-    cy: 50,
-    label: false,
-    labelPosition: 50,
-    lengthAngle: 360,
-    lineWidth: 100,
-    onMouseOver: undefined,
-    onMouseOut: undefined,
-    onClick: undefined,
-    paddingAngle: 0,
-    radius: 50,
-    rounded: false,
-    startAngle: 0,
-    viewBoxSize: [100, 100],
-  };
-  static propTypes = {
-    animate: PropTypes.bool,
-    animationDuration: PropTypes.number,
-    animationEasing: PropTypes.string,
-    background: PropTypes.string,
-    children: PropTypes.node,
-    className: PropTypes.string,
-    cx: PropTypes.number,
-    cy: PropTypes.number,
-    data: dataPropType,
-    injectSvg: PropTypes.func,
-    label: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.element,
-      PropTypes.bool,
-    ]),
-    labelPosition: PropTypes.number,
-    labelStyle: stylePropType,
-    lengthAngle: PropTypes.number,
-    lineWidth: PropTypes.number,
-    onClick: PropTypes.func,
-    onMouseOut: PropTypes.func,
-    onMouseOver: PropTypes.func,
-    paddingAngle: PropTypes.number,
-    radius: PropTypes.number,
-    reveal: PropTypes.number,
-    rounded: PropTypes.bool,
-    segmentsStyle: stylePropType,
-    startAngle: PropTypes.number,
-    style: stylePropType,
-    totalValue: PropTypes.number,
-    viewBoxSize: PropTypes.arrayOf(PropTypes.number),
-  };
-
-  hideSegments: boolean;
-  initialAnimationTimerId: null | number;
-  initialAnimationRAFId: null | number;
-
-  constructor(props: Props) {
-    super(props);
-
-    if (props.animate === true) {
-      this.hideSegments = true;
-    }
-  }
-
-  componentDidMount() {
-    if (this.props.animate === true && requestAnimationFrame) {
-      this.initialAnimationTimerId = setTimeout(() => {
-        this.initialAnimationTimerId = null;
-        this.initialAnimationRAFId = requestAnimationFrame(() => {
-          this.initialAnimationRAFId = null;
-          this.startAnimation();
-        });
-      });
-    }
-  }
-
-  componentWillUnmount() {
-    if (this.initialAnimationTimerId) {
-      clearTimeout(this.initialAnimationTimerId);
-    }
-    if (this.initialAnimationRAFId) {
-      cancelAnimationFrame(this.initialAnimationRAFId);
-    }
-  }
-
-  startAnimation() {
-    this.hideSegments = false;
-    this.forceUpdate();
-  }
-
-  render() {
-    const props = this.props;
-    if (props.data === undefined) {
-      return null;
-    }
-    const extendedData = extendData(props);
-
-    return (
-      <div className={props.className} style={props.style}>
-        <svg
-          viewBox={`0 0 ${props.viewBoxSize[0]} ${props.viewBoxSize[1]}`}
-          width="100%"
-          height="100%"
-          style={{ display: 'block' }}
-        >
-          {renderSegments(extendedData, props, this.hideSegments)}
-          {props.label && renderLabels(extendedData, props)}
-          {props.injectSvg && props.injectSvg()}
-        </svg>
-        {props.children}
-      </div>
-    );
-  }
 }
