+++ src/Chart/renderSegments.tsx
@@ -1,66 +1,8 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React from 'react';
 import Path from '../Path';
-import { default as DefaultLabel, Props as LabelProps } from '../Label';
-import {
-  degreesToRadians,
-  evaluateLabelTextAnchor,
-  extractPercentage,
-  valueBetween,
-} from '../utils';
-import { dataPropType, stylePropType } from '../propTypes';
-import {
-  Data,
-  ExtendedData,
-  EventHandler,
-  LabelProp,
-  StyleObject,
-} from '../commonTypes';
-
-function extractAbsoluteCoordinates(props: Props) {
-  const [viewBoxWidth, viewBoxHeight] = props.viewBoxSize;
-  return {
-    cx: extractPercentage(props.cx, viewBoxWidth),
-    cy: extractPercentage(props.cy, viewBoxHeight),
-    radius: extractPercentage(props.radius, viewBoxWidth),
-  };
-}
-
-function sumValues(data: Data) {
-  return data.reduce((acc, dataEntry) => acc + dataEntry.value, 0);
-}
-
-// Append "percentage", "degrees" and "startOffset" into each data entry
-function extendData({
-  data,
-  lengthAngle: totalAngle,
-  totalValue,
-  paddingAngle,
-}: Props) {
-  const total = totalValue || sumValues(data);
-  const normalizedTotalAngle = valueBetween(totalAngle, -360, 360);
-  const numberOfPaddings =
-    Math.abs(normalizedTotalAngle) === 360 ? data.length : data.length - 1;
-  const singlePaddingDegrees = Math.abs(paddingAngle) * Math.sign(totalAngle);
-  const degreesTakenByPadding = singlePaddingDegrees * numberOfPaddings;
-  const degreesTakenByPaths = normalizedTotalAngle - degreesTakenByPadding;
-  let lastSegmentEnd = 0;
-
-  // @NOTE: Shall we evaluate percentage accordingly to dataEntry.value's sign?
-  return data.map(dataEntry => {
-    const valueInPercentage = total === 0 ? 0 : (dataEntry.value / total) * 100;
-    const degrees = extractPercentage(degreesTakenByPaths, valueInPercentage);
-    const startOffset = lastSegmentEnd;
-    lastSegmentEnd = lastSegmentEnd + degrees + singlePaddingDegrees;
-
-    return {
-      percentage: valueInPercentage,
-      degrees,
-      startOffset,
-      ...dataEntry,
-    };
-  });
-}
+import { extractPercentage, extractAbsoluteCoordinates } from '../utils';
+import { ExtendedData, StyleObject } from '../commonTypes';
+import { Props as ChartProps } from './index';
 
 function makeSegmentTransitionStyle(
   duration: number,
@@ -80,60 +22,11 @@ function makeSegmentTransitionStyle(
   };
 }
 
-function renderLabelItem(
-  option: LabelProp,
-  labelProps: LabelProps,
-  value: number
+export default function renderSegments(
+  data: ExtendedData,
+  props: ChartProps,
+  hide: Boolean
 ) {
-  if (React.isValidElement(option)) {
-    return React.cloneElement(option, labelProps);
-  }
-
-  let label: number | string | React.ReactElement = value;
-  if (typeof option === 'function') {
-    label = option(labelProps);
-    if (React.isValidElement(label)) {
-      return label;
-    }
-  }
-
-  return <DefaultLabel {...labelProps}>{label}</DefaultLabel>;
-}
-
-function renderLabels(data: ExtendedData, props: Props) {
-  const { cx, cy, radius } = extractAbsoluteCoordinates(props);
-  const labelPosition = extractPercentage(radius, props.labelPosition);
-
-  return data.map((dataEntry, index) => {
-    const startAngle = props.startAngle + dataEntry.startOffset;
-    const halfAngle = startAngle + dataEntry.degrees / 2;
-    const halfAngleRadians = degreesToRadians(halfAngle);
-    const dx = Math.cos(halfAngleRadians) * labelPosition;
-    const dy = Math.sin(halfAngleRadians) * labelPosition;
-
-    // This object is passed as props to the "label" component
-    const labelProps = {
-      key: `label-${dataEntry.key || index}`,
-      x: cx,
-      y: cy,
-      dx,
-      dy,
-      textAnchor: evaluateLabelTextAnchor({
-        lineWidth: props.lineWidth,
-        labelPosition: props.labelPosition,
-        labelHorizontalShift: dx,
-      }),
-      data: data,
-      dataIndex: index,
-      color: dataEntry.color,
-      style: props.labelStyle,
-    };
-
-    return renderLabelItem(props.label, labelProps, dataEntry.value);
-  });
-}
-
-function renderSegments(data: ExtendedData, props: Props, hide: Boolean) {
   let style = props.segmentsStyle;
   if (props.animate) {
     const transitionStyle = makeSegmentTransitionStyle(
@@ -211,151 +104,4 @@ function renderSegments(data: ExtendedData, props: Props, hide: Boolean) {
   }
 
   return paths;
-}
-
-type Props = typeof ReactMinimalPieChart.defaultProps & {
-  animate?: boolean;
-  animationDuration?: number;
-  animationEasing?: string;
-  background?: string;
-  className?: string;
-  cx?: number;
-  cy?: number;
-  data: Data;
-  lengthAngle?: number;
-  lineWidth?: number;
-  injectSvg?: () => React.ReactElement | void;
-  label?: LabelProp;
-  labelPosition?: number;
-  labelStyle?: StyleObject;
-  onClick?: EventHandler;
-  onMouseOut?: EventHandler;
-  onMouseOver?: EventHandler;
-  paddingAngle?: number;
-  radius?: number;
-  reveal?: number;
-  rounded?: boolean;
-  segmentsStyle?: StyleObject;
-  startAngle?: number;
-  style?: StyleObject;
-  totalValue?: number;
-  viewBoxSize?: [number, number];
-};
-
-export default class ReactMinimalPieChart extends Component<Props> {
-  static displayName = 'ReactMinimalPieChart';
-  static defaultProps = {
-    animate: false,
-    animationDuration: 500,
-    animationEasing: 'ease-out',
-    cx: 50,
-    cy: 50,
-    label: false,
-    labelPosition: 50,
-    lengthAngle: 360,
-    lineWidth: 100,
-    onMouseOver: undefined,
-    onMouseOut: undefined,
-    onClick: undefined,
-    paddingAngle: 0,
-    radius: 50,
-    rounded: false,
-    startAngle: 0,
-    viewBoxSize: [100, 100],
-  };
-  static propTypes = {
-    animate: PropTypes.bool,
-    animationDuration: PropTypes.number,
-    animationEasing: PropTypes.string,
-    background: PropTypes.string,
-    children: PropTypes.node,
-    className: PropTypes.string,
-    cx: PropTypes.number,
-    cy: PropTypes.number,
-    data: dataPropType,
-    injectSvg: PropTypes.func,
-    label: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.element,
-      PropTypes.bool,
-    ]),
-    labelPosition: PropTypes.number,
-    labelStyle: stylePropType,
-    lengthAngle: PropTypes.number,
-    lineWidth: PropTypes.number,
-    onClick: PropTypes.func,
-    onMouseOut: PropTypes.func,
-    onMouseOver: PropTypes.func,
-    paddingAngle: PropTypes.number,
-    radius: PropTypes.number,
-    reveal: PropTypes.number,
-    rounded: PropTypes.bool,
-    segmentsStyle: stylePropType,
-    startAngle: PropTypes.number,
-    style: stylePropType,
-    totalValue: PropTypes.number,
-    viewBoxSize: PropTypes.arrayOf(PropTypes.number),
-  };
-
-  hideSegments: boolean;
-  initialAnimationTimerId: null | number;
-  initialAnimationRAFId: null | number;
-
-  constructor(props: Props) {
-    super(props);
-
-    if (props.animate === true) {
-      this.hideSegments = true;
-    }
-  }
-
-  componentDidMount() {
-    if (this.props.animate === true && requestAnimationFrame) {
-      this.initialAnimationTimerId = setTimeout(() => {
-        this.initialAnimationTimerId = null;
-        this.initialAnimationRAFId = requestAnimationFrame(() => {
-          this.initialAnimationRAFId = null;
-          this.startAnimation();
-        });
-      });
-    }
-  }
-
-  componentWillUnmount() {
-    if (this.initialAnimationTimerId) {
-      clearTimeout(this.initialAnimationTimerId);
-    }
-    if (this.initialAnimationRAFId) {
-      cancelAnimationFrame(this.initialAnimationRAFId);
-    }
-  }
-
-  startAnimation() {
-    this.hideSegments = false;
-    this.forceUpdate();
-  }
-
-  render() {
-    const props = this.props;
-    if (props.data === undefined) {
-      return null;
-    }
-    const extendedData = extendData(props);
-
-    return (
-      <div className={props.className} style={props.style}>
-        <svg
-          viewBox={`0 0 ${props.viewBoxSize[0]} ${props.viewBoxSize[1]}`}
-          width="100%"
-          height="100%"
-          style={{ display: 'block' }}
-        >
-          {renderSegments(extendedData, props, this.hideSegments)}
-          {props.label && renderLabels(extendedData, props)}
-          {props.injectSvg && props.injectSvg()}
-        </svg>
-        {props.children}
-      </div>
-    );
-  }
 }
